INTRODUCED SHARED POINTERS

+++ src/Address.h
@@ -11,13 +11,12 @@ class Address{
 		uint_fast8_t pin;
 		bool sel[6]; //module slot selectors
 	public:
-		uint_fast8_t _id;
+		unsigned int _id;
 		Address(uint_fast8_t);
 		Address(const Address&);
 		bool getSel(uint_fast8_t i) const{return sel[i];};
 		void setForReading() const;
 		virtual void setForWriting() const{};
-	//	uint_fast8_t readIdMux() const;
 		uint_fast8_t getPin() const{return this->pin;};
 		uint_fast8_t toInt()const{return _id;};
 };
